Remove getSlaveIndex from MotorController

+++ march_hardware/include/march_hardware/ethercat/slave.h
@@ -31,7 +31,7 @@ public:
 
   ~Slave() noexcept override = default;
 
-  int getSlaveIndex() const
+  uint16_t getSlaveIndex() const
   {
     return this->slave_index_;
   }
